==> Code/step2_extract_complexity_emotion_and_stylistic_features.ipynb <==
author: "Anu Shrestha"

# Extract and save features (Complexity, stylistic and  emotion features)
This file contains the code that extracts and save features (Complexity, stylistic and  emotion features).
This function add (rewrite over the same csv file) all extracted features.



!pip install ReadabilityCalculator
!pip install textstat
!pip install liwc
!pip install vaderSentiment

import textstat
from readcalc import readcalc
import nltk
nltk.download("punkt")
nltk.download('stopwords')
nltk.download('wordnet')
nltk.download('punkt')
nltk.download('averaged_perceptron_tagger')
nltk.download('wordnet')
from sklearn.feature_extraction.text import ENGLISH_STOP_WORDS
from spacy.lang.en import STOP_WORDS
from nltk.corpus import stopwords
stops = list(set(stopwords.words('english') + list(set(ENGLISH_STOP_WORDS)) + list(set(STOP_WORDS)) + ["http"]))
from nltk.stem.wordnet import WordNetLemmatizer
lmtzr = WordNetLemmatizer()

from nltk.corpus import wordnet
from nltk.tokenize import word_tokenize
import re
import string
from nltk.tag import pos_tag
from nltk.tokenize import RegexpTokenizer
from nltk import sent_tokenize
import liwc
import os

from textblob import TextBlob
import pandas as pd
import numpy as np
from collections import Counter
from tqdm import tqdm
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer


from google.colab import drive
drive.mount('/content/gdrive')

## Import external python file

import sys
sys.path.append('/content/gdrive/My Drive/ECIR 2021 Reproducibility/Code/code_to_publish')

import readability
import stylistic_features
import emotion_features

def extract_complexity_emotion_stylistic_features(path, filename, modality_type):
  '''
  Extract all the features used in paper for text in input filename.
  '''
  df = pd.read_csv(os.path.join(path, filename+".csv"))
  df = df.sample(frac=0.01)
  
  #Extract complexity features
  df = readability.compute_readability(df, 'news_'+modality_type)
  df = readability.compute_syntactic(df, 'news_'+modality_type)
  df['lexical_diversity'] = df['news_'+modality_type].apply(stylistic_features.lexical_diversity)
  df['wlen'] = df['news_'+modality_type].apply(stylistic_features.average_word_length)

  # Extract stylistic features
  df = stylistic_features.part_of_speech(df, 'news_'+modality_type)
  df = stylistic_features.numeric_features(df, 'news_'+modality_type)

  # Extract emotion features
  emo_dic_path = "/content/gdrive/My Drive/Emotion_vectors_from_text/emotion_itensity.txt"# path to the emotion intensity lexicon dictionary
  df = emotion_features.emotion_NRC(df, emo_dic_path, 'news_'+modality_type )
  df = emotion_features.sentiment_strength_vader(df, 'news_'+modality_type)

  # Save extracted features
  if not os.path.exists(os.path.join(path,'Generated_features')):
    os.makedirs(os.path.join(path,'Generated_features'))

  df.to_pickle(os.path.join(path,"Generated_features/features_for_"+filename+"_noLIWC.pkl"))
  print("Saving extracted features completed!")



path = '/content/gdrive/My Drive/ECIR 2021 Reproducibility/Data/' #path to the folder containing datasets
title_filenames = ['title_politifact','title_buzzfeed','title_gossipcop'] # change filenames as per your requirement
text_filenames = ['text_politifact','text_buzzfeed','text_gossipcop'] # change filenames as per your requirement

for title_filename in title_filenames:
  extract_complexity_emotion_stylistic_features(path, title_filename, modality_type='title')

for text_filename in text_filenames:
  extract_complexity_emotion_stylistic_features(path, text_filename, modality_type='text')
