==> Code/step4_statistical_test.ipynb <==
author: "Anu Shrestha"

# Run statistical t test 

This code will run statistical t test to compare which features differs between fake and real news as shown in table 2 and 3 in paper.

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')
from google.colab import drive
drive.mount('/content/gdrive')

import sys
sys.path.append('/content/gdrive/My Drive/ECIR 2021 Reproducibility/Code/code_to_publish')

import statistical_tests

df_title_gossipcop = pd.read_pickle("/content/gdrive/My Drive/ECIR 2021 Reproducibility/Data/Generated_features/all_features_for_title_gossipcop.pkl")
df_title_politifact= pd.read_pickle("/content/gdrive/My Drive/ECIR 2021 Reproducibility/Data/Generated_features/all_features_for_title_politifact.pkl")
df_text_gossipcop= pd.read_pickle("/content/gdrive/My Drive/ECIR 2021 Reproducibility/Data/Generated_features/all_features_for_text_gossipcop.pkl")
df_text_politifact= pd.read_pickle("/content/gdrive/My Drive/ECIR 2021 Reproducibility/Data/Generated_features/all_features_for_text_politifact.pkl")
df_title_buzzfeed= pd.read_pickle("/content/gdrive/My Drive/ECIR 2021 Reproducibility/Data/Generated_features/all_features_for_title_buzzfeed.pkl")
df_text_buzzfeed= pd.read_pickle("/content/gdrive/My Drive/ECIR 2021 Reproducibility/Data/Generated_features/all_features_for_text_buzzfeed.pkl")

# Statistical t test

stat_title_g = statistical_tests.t_test(df_title_gossipcop)
stat_title_p = statistical_tests.t_test(df_title_politifact)
stat_title_b = statistical_tests.t_test(df_title_buzzfeed)
stat_text_g = statistical_tests.t_test(df_text_gossipcop)
stat_text_p = statistical_tests.t_test(df_text_politifact)
stat_text_b = statistical_tests.t_test(df_text_buzzfeed)

stat_title_g

## Save result of statistical test for only those features that are used in paper

features_used_in_paper = ['Analytic','insight','cause','discrep','tentat','certain','differ','affiliation','power','reward','risk','work','leisure',
                            'money','relig','Tone','affect','WC','WPS','num_nouns','num_propernouns','num_personalnouns','num_ppssessivenouns',
                            'num_whpronoun','num_determinants','num_whdeterminants','num_cnum','num_adverb','num_interjections','num_verb','num_adj',
                            'num_vbd','num_vbg','num_vbn','num_vbp','num_vbz','focuspast','focusfuture','i','we','you','shehe','quant','compare','Exclam',
                            'negate','swear','netspeak','interrog','count_uppercased','percentage_stopwords','AllPunc','Quote', 'lexical_diversity','wlen',
                            'gunning_fog_index','smog_index','flesch_kincaid_grade_level',
                          'Anger','Anticipation','Disgust','Fear','Joy', 'Sadness', 'Surprise', 'Trust','neg','pos','posemo','negemo','anx']

def num_sig(df):
  ddf_sig = df[df.feature.isin(features_used_in_paper)]
  return ddf_sig

df_text_b = num_sig(stat_text_b)
df_title_b = num_sig(stat_title_b)
df_title_g = num_sig(stat_title_g)
df_text_g = num_sig(stat_text_g)
df_text_p = num_sig(stat_text_p)
df_title_p = num_sig(stat_title_p)

df_text_b

# Arrange all the test result to a single table

title = set(df_title_p.feature).union(set(df_title_g.feature)).union(set(df_title_b.feature))

text = set(df_text_p.feature).union(set(df_text_g.feature)).union(set(df_text_b.feature))

df_g = df_text_g[df_text_g.feature.isin(text)]
df_p = df_text_p[df_text_p.feature.isin(text)]
df_b = df_text_b[df_text_b.feature.isin(text)]

df_g.head()

df_b.rename(columns={'feature_differ':'BuzzFeed'}, inplace=True)
df_g.rename(columns={'feature_differ':'GossipCop'}, inplace=True)
df_p.rename(columns={'feature_differ':'PolitiFact'}, inplace=True)

df_p_b = df_p[['feature','PolitiFact']].merge(df_b[['feature','BuzzFeed']], on='feature',how='outer')

df_p_b_g = df_p_b.merge(df_g[['feature','GossipCop']], on='feature',how='outer')

df_p_b_g.PolitiFact = df_p_b_g.PolitiFact.apply(lambda x: 'F>R' if x=='Fake > Real' else ('R>F' if x=='Real > Fake' else ''))
df_p_b_g.BuzzFeed = df_p_b_g.BuzzFeed.apply(lambda x: 'F>R' if x=='Fake > Real' else ('R>F' if x=='Real > Fake' else ''))
df_p_b_g.GossipCop = df_p_b_g.GossipCop.apply(lambda x: 'F>R' if x=='Fake > Real' else ('R>F' if x=='Real > Fake' else ''))

df_p_b_g
